TST/REF: collect Index.equals tests (#38011)
